+++ org.eclim.jdt/java/org/eclim/plugin/jdt/command/delegate/DelegateCommand.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2005 - 2011  Eric Van Dewoestine
+ * Copyright (C) 2005 - 2012  Eric Van Dewoestine
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -47,6 +47,8 @@ import org.eclipse.jdt.core.IMethod;
 import org.eclipse.jdt.core.IType;
 import org.eclipse.jdt.core.ITypeParameter;
 import org.eclipse.jdt.core.Signature;
+
+import org.eclipse.jdt.core.formatter.CodeFormatter;
 
 /**
  * Command to handle creation of delegate methods.
@@ -219,6 +221,9 @@ public class DelegateCommand
     String result = TemplateUtils.evaluate(resources, TEMPLATE, values);
     Position position = TypeUtils.getPosition(type,
         type.createMethod(result, sibling, false, null));
+    JavaUtils.format(
+        src, CodeFormatter.K_COMPILATION_UNIT,
+        position.getOffset(), position.getLength());
 
     return position;
   }